refactor: type/width fidelity + faithful comments vs Java rev-225 (no behavior change)
Closes the latent-divergence and comment items from the 2026-05-28 parity audit.
All changes are behavior-preserving for the values that actually occur.

- world.go + client.LevelTileFlags: re-type the six level map fields
  byte -> int8 to match Java's signed byte[][][], so widening with int(...)
  sign-extends like Java; the Build() shademap >> is promoted per-term via
  int(...) (Java byte->int promotion). Stored values stay 0..50/IDs, so results
  are identical.
- projectileentity: var6 -> float64 (Java double, so var6*var6 is a double mul).
- playerentity.getSequencedModel: truncate the <<8/<<16 cache-key terms to int32
  (Java int wrap). Internal LruCache key only; never observed/sent.
- model / signlink: emit Java-verbatim diagnostics ("Error model:%d not found\!",
  "Error: ") per the faithful-System.out.println convention.
- signlink.GetHash: iterate runes with Java trim() semantics (chars <= U+0020);
  removes a byte-vs-UTF16 trap (dead code — no callers).
- comments: SetLowMem/SetHighMem // Java: rename refs; SetMidi null<->"" rationale;
  pix3d palette-index sign-extension note (Go intentionally cannot crash where
  Java would); viewbox.go header corrected (Go emulates Java's standalone
  frame \!= null case, via the platform seam — not the applet case, not Gio).

+++ pkg/jagex2/dash3d/entity/playerentity/playerentity.go
@@ -203,11 +203,14 @@ func (e *PlayerEntity) GetSequencedModel() *model.Model {
 		}
 		if var8.RightHand >= 0 {
 			var6 = var8.RightHand
-			var2 += int64((var6 - e.Appearances[5]) << 8)
+			// Java: `var2 += var6 - appearances[5] << 8` is 32-bit int arithmetic,
+			// sign-extended into the long var2. int32(...) reproduces that wrap; Go's
+			// 64-bit int would otherwise diverge for high righthand/lefthand values.
+			var2 += int64(int32((var6 - e.Appearances[5]) << 8))
 		}
 		if var8.LeftHand >= 0 {
 			var7 = var8.LeftHand
-			var2 += int64((var7 - e.Appearances[3]) << 16)
+			var2 += int64(int32((var7 - e.Appearances[3]) << 16))
 		}
 	} else if e.SecondarySeqID >= 0 {
 		var4 = seqtype.Instances[e.SecondarySeqID].Frames[e.SecondarySeqFrame]
